# Conflicts:
#	ES Game Library/Content/FX/HDR.fx
#	ES Game Library/GameMain.cpp
#	ES Game Library/GameMain.h
#	ES Game Library/MakeClasses/Fujimura/BgmSingleton.cpp
#	ES Game Library/MakeClasses/Fujimura/UI/BackMovie.cpp

+++ ES Game Library/MakeClasses/Fujimura/Judgement/EventNoteJudgement.h
@@ -11,7 +11,7 @@ public:
 	EventNoteJudgement();
 	~EventNoteJudgement();
 
-	virtual JUDGE Judge(Note* judgeNote, DWORD nowTime, RawInputMouse& mouse);
+	virtual JUDGE Judge(Note* judgeNote, LONG nowTime, RawInputMouse& mouse);
 
 private:
