+++ ES Game Library/MakeClasses/Fujimura/Note/DrawNote/LongNoteDraw.h
@@ -8,6 +8,6 @@ public:
 	LongNoteDraw();
 	~LongNoteDraw();
 
-	virtual void NoteDraw(Note* note, BaseLane* lane, DWORD nowTime, DWORD drawRangeTime);
+	virtual void NoteDraw(Note* note, BaseLane* lane, LONG nowTime, LONG drawRangeTime);
 
 };
